feat(input): use input primitive
