+++ packages/components/input/src/Input.tsx
@@ -1,48 +1,27 @@
 import { useFormFieldState } from '@spark-ui/form-field'
-import { ComponentPropsWithoutRef, forwardRef } from 'react'
+import { forwardRef } from 'react'
 
 import { inputStyles, InputStylesProps } from './Input.styles'
 import { useInputGroup } from './InputGroupContext'
+import { InputPrimitive, InputPrimitiveProps } from './InputPrimitive'
 
-export interface InputProps
-  extends ComponentPropsWithoutRef<'input'>,
-    Omit<
-      InputStylesProps,
-      | 'isDisabled'
-      | 'isLeftAddonVisible'
-      | 'isRightAddonVisible'
-      | 'isLeftElementVisible'
-      | 'isRightElementVisible'
-    > {}
+export interface InputProps extends InputPrimitiveProps, InputStylesProps {}
 
 export const Input = forwardRef<HTMLInputElement, InputProps>(
-  ({ className, intent: intentProp = 'neutral', disabled: disabledProp, ...others }, ref) => {
+  (
+    { className: classNameProp, intent: intentProp = 'neutral', disabled: disabledProp, ...others },
+    ref
+  ) => {
     const field = useFormFieldState()
-    const group = useInputGroup() || {}
-
-    const { isLeftAddonVisible, isRightAddonVisible, isLeftElementVisible, isRightElementVisible } =
-      group
-    const { id, name, isInvalid, isRequired, description } = field
-    const intent = isInvalid && intentProp !== 'none' ? 'error' : intentProp
-    const isDisabled = group.isDisabled ?? disabledProp
+    const group = useInputGroup()
+    const { isInvalid } = field
+    const isGrouped = !!group
+    const intent = isInvalid ? 'error' : intentProp
 
     return (
-      <input
+      <InputPrimitive
         ref={ref}
-        id={id}
-        name={name}
-        className={inputStyles({
-          className,
-          intent,
-          isLeftAddonVisible: !!isLeftAddonVisible,
-          isRightAddonVisible: !!isRightAddonVisible,
-          isLeftElementVisible: !!isLeftElementVisible,
-          isRightElementVisible: !!isRightElementVisible,
-        })}
-        disabled={isDisabled}
-        required={isRequired}
-        aria-describedby={description}
-        aria-invalid={isInvalid}
+        className={inputStyles({ className: classNameProp, intent, isGrouped })}
         {...others}
       />
     )
