+++ packages/components/input/src/InputGroup.styles.tsx
@@ -1,5 +1,5 @@
 import { cva, VariantProps } from 'class-variance-authority'
 
-export const inputGroupStyles = cva(['relative', 'inline-flex'])
+export const inputGroupStyles = cva(['relative', 'inline-flex', 'w-full'])
 
 export type InputGroupStylesProps = VariantProps<typeof inputGroupStyles>
