+++ packages/components/input/src/index.ts
@@ -9,6 +9,7 @@ import { InputRightElement, InputRightElementProps } from './InputRightElement'
 export { useInputGroup } from './InputGroupContext'
 
 export * from './Input'
+export * from './InputPrimitive'
 export * from './InputContainer'
 export { type InputGroupProps } from './InputGroup'
 export { type InputLeftAddonProps } from './InputLeftAddon'
